revert Contains tests
